Update Duke to match previous changes
